feat: Refactor & Initial commit
Initialize this project with some basic structures.

This is a project that named Snake, trying to build a neural network in C.
Its originally write in wsl2,
now I will refactor it and build all of the code in windows vscode.
Maybe in the future, I will learn some cuda and try to get some gpu features,
although for now it's a pure cpu project.

- Add CMakeLists.txt for build configuration.
- Add initial Tensor library source and header files.
- Add basic test structure for tensor shape.
- Add .gitignore, LICENSE, and README.md.
- Finish shape logics basically.